+++ src/analysis/data_handler.rs
@@ -8,10 +8,10 @@ use std::io::{BufRead, BufReader, Seek};
 //Stores a datafile and the corresponding test data
 pub struct DataHandler {
     file: File,
-    timestamps: Vec<f32>,
-    trajectory: Vec<[f32; 3]>,
-    ori: Vec<[f32; 3]>,
-    forces: Vec<[f32; 6]>,
+    timestamps: Vec<f64>,
+    trajectory: Vec<[f64; 3]>,
+    ori: Vec<[f64; 3]>,
+    forces: Vec<[f64; 6]>,
 }
 
 impl DataHandler {
@@ -21,10 +21,10 @@ impl DataHandler {
         let mut file = File::open(filepath)?;
 
         //Create the empty vectors
-        let mut timestamps: Vec<f32> = vec![];
-        let mut trajectory: Vec<[f32; 3]> = vec![];
-        let mut ori: Vec<[f32; 3]> = vec![];
-        let mut forces: Vec<[f32; 6]> = vec![];
+        let mut timestamps: Vec<f64> = vec![];
+        let mut trajectory: Vec<[f64; 3]> = vec![];
+        let mut ori: Vec<[f64; 3]> = vec![];
+        let mut forces: Vec<[f64; 6]> = vec![];
 
         //Go line by line through the file
         for line in BufReader::new(file.try_clone()?).lines() {
@@ -72,7 +72,7 @@ impl DataHandler {
 
     //Function which gets the rectangular bounds of a trajectory
     //Z pos currently unused
-    pub fn get_traj_rect_bnds(&mut self) -> Result<[f32; 4], anyhow::Error> {
+    pub fn get_traj_rect_bnds(&mut self) -> Result<[f64; 4], anyhow::Error> {
         let mut min_x = 9999.0;
         let mut max_x = -9999.0;
         let mut min_y = 9999.0;
@@ -98,19 +98,19 @@ impl DataHandler {
     }
 
     //Extracts every trajectory point from the datafile
-    pub fn get_traj(&self) -> Vec<[f32; 3]> {
+    pub fn get_traj(&self) -> Vec<[f64; 3]> {
         //Return the trajectory
         self.trajectory.clone()
     }
 
     //Returns the trajectory and force data for the file - in an vector comprised of a tuple trajectory/force vector pairs
-    pub fn get_force(&self) -> Vec<[f32; 6]> {
+    pub fn get_force(&self) -> Vec<[f64; 6]> {
         self.forces.clone()
     }
 
     //Returns a tuple which pairs trajectory and force data together (essentially a zip)
-    pub fn get_traj_force_pairs(&mut self) -> Vec<([f32; 3], [f32; 6])> {
-        let mut traj_force_pairs: Vec<([f32; 3], [f32; 6])> = vec![];
+    pub fn get_traj_force_pairs(&mut self) -> Vec<([f64; 3], [f64; 6])> {
+        let mut traj_force_pairs: Vec<([f64; 3], [f64; 6])> = vec![];
 
         //Go through every point in the trajectory, extract it and pair with the force data
         for (i, pnt) in self.trajectory.iter_mut().enumerate() {
@@ -122,8 +122,8 @@ impl DataHandler {
 }
 
 //turns a string in the format "[x,y,z]" into a vector
-fn str_to_traj_ori(x: &str) -> Result<[f32; 3], anyhow::Error> {
-    let mut curr_vec: [f32; 3] = [f32::NAN, f32::NAN, f32::NAN];
+fn str_to_traj_ori(x: &str) -> Result<[f64; 3], anyhow::Error> {
+    let mut curr_vec: [f64; 3] = [f64::NAN, f64::NAN, f64::NAN];
 
     //Strip the "[" and "]"
     let x_strip = string_tools::rem_first_and_last(x);
@@ -142,8 +142,8 @@ fn str_to_traj_ori(x: &str) -> Result<[f32; 3], anyhow::Error> {
 }
 
 //Parses the force string into the force vector
-fn str_to_force(x: &str) -> Result<[f32; 6], anyhow::Error> {
-    let mut curr_vec: [f32; 6] = [f32::NAN, f32::NAN, f32::NAN, f32::NAN, f32::NAN, f32::NAN];
+fn str_to_force(x: &str) -> Result<[f64; 6], anyhow::Error> {
+    let mut curr_vec: [f64; 6] = [f64::NAN, f64::NAN, f64::NAN, f64::NAN, f64::NAN, f64::NAN];
 
     //Strip the "[" and "]"
     let x_strip = string_tools::rem_first_and_last(x);
