+++ src/config.rs
@@ -21,20 +21,20 @@ pub struct Config {
 
 #[derive(Debug)]
 pub struct CamInfo {
-    rel_pos: [f32; 3],
-    rel_ori: [f32; 3],
-    x_scale: f32,
-    y_scale: f32,
+    rel_pos: [f64; 3],
+    rel_ori: [f64; 3],
+    x_scale: f64,
+    y_scale: f64,
 }
 
 #[derive(Debug)]
 pub struct RobInfo {
     rob_name: String,
     //Position and orientation information required for transformation to global world frame (0,0 in terrian box)
-    pos_for_zero: [f32; 3],
-    ori_for_zero: [f32; 3],
+    pos_for_zero: [f64; 3],
+    ori_for_zero: [f64; 3],
     //Height that the robot registers where the end effector sits minimally above the soil
-    min_embed_height: f32,
+    min_embed_height: f64,
 }
 
 impl RobInfo {}
@@ -62,7 +62,7 @@ impl Default for CamInfo {
         CamInfo {
             rel_pos: [250.0, 250.0, 250.0],
             //around 45 degrees facing downward
-            rel_ori: [0.785, std::f32::consts::PI, 0.0],
+            rel_ori: [0.785, std::f64::consts::PI, 0.0],
             //Scale from mm to m
             x_scale: 0.001,
             y_scale: 0.001,
@@ -134,10 +134,10 @@ impl Config {
 
 impl CamInfo {
     pub fn create_cam_info(
-        rel_pos: [f32; 3],
-        rel_ori: [f32; 3],
-        x_scale: f32,
-        y_scale: f32,
+        rel_pos: [f64; 3],
+        rel_ori: [f64; 3],
+        x_scale: f64,
+        y_scale: f64,
     ) -> Self {
         Self {
             rel_pos,
@@ -153,10 +153,10 @@ impl CamInfo {
         let cam_inf_split = cam_info_line.split("[");
         let mut ind_cnt = 0;
 
-        let mut rel_pos = [f32::NAN, f32::NAN, f32::NAN];
-        let mut rel_ori = [f32::NAN, f32::NAN, f32::NAN];
-        let mut x_scale = f32::NAN;
-        let mut y_scale = f32::NAN;
+        let mut rel_pos = [f64::NAN, f64::NAN, f64::NAN];
+        let mut rel_ori = [f64::NAN, f64::NAN, f64::NAN];
+        let mut x_scale = f64::NAN;
+        let mut y_scale = f64::NAN;
 
         for split in cam_inf_split {
             //Split again to isolate the data
@@ -201,10 +201,10 @@ impl CamInfo {
         let cam_config_file = File::open(fp)?;
 
         //Have the default values initialised - incase they aren't overwritten
-        let mut rel_pos: [f32; 3] = [250.0, 250.0, 250.0];
-        let mut rel_ori: [f32; 3] = [0.785, std::f32::consts::PI, 0.0];
-        let mut x_scale: f32 = 0.001;
-        let mut y_scale: f32 = 0.001;
+        let mut rel_pos: [f64; 3] = [250.0, 250.0, 250.0];
+        let mut rel_ori: [f64; 3] = [0.785, std::f64::consts::PI, 0.0];
+        let mut x_scale: f64 = 0.001;
+        let mut y_scale: f64 = 0.001;
 
         //Go through each line and parse the info
         for line in BufReader::new(cam_config_file).lines() {
@@ -231,7 +231,7 @@ impl CamInfo {
         })
     }
 
-    fn extract_scale(line: String) -> Result<f32, anyhow::Error> {
+    fn extract_scale(line: String) -> Result<f64, anyhow::Error> {
         //Access the value
         let line_split: Vec<&str> = line.split("[").collect();
         let val = line_split[1].replace("]", "");
@@ -240,18 +240,18 @@ impl CamInfo {
         Ok(val.parse()?)
     }
 
-    pub fn rel_pos(&self) -> [f32; 3] {
+    pub fn rel_pos(&self) -> [f64; 3] {
         self.rel_pos
     }
 
-    pub fn rel_ori(&self) -> [f32; 3] {
+    pub fn rel_ori(&self) -> [f64; 3] {
         self.rel_ori
     }
 
-    pub fn x_scale(&self) -> f32 {
+    pub fn x_scale(&self) -> f64 {
         self.x_scale
     }
-    pub fn y_scale(&self) -> f32 {
+    pub fn y_scale(&self) -> f64 {
         self.y_scale
     }
 }
@@ -306,7 +306,7 @@ impl RobInfo {
         let para_split: Vec<&str> = name_split[2].split("[").collect();
 
         //Access the pos
-        let mut pos_for_zero = [f32::NAN, f32::NAN, f32::NAN];
+        let mut pos_for_zero = [f64::NAN, f64::NAN, f64::NAN];
         let poses = para_split[1].replace("] ORI:", "");
         let poses: Vec<&str> = poses.split(",").collect();
         for (pos_cnt, pos) in poses.into_iter().enumerate() {
@@ -314,7 +314,7 @@ impl RobInfo {
         }
 
         //Access the orientation
-        let mut ori_for_zero = [f32::NAN, f32::NAN, f32::NAN];
+        let mut ori_for_zero = [f64::NAN, f64::NAN, f64::NAN];
         let oris = para_split[2].replace("]", "");
         let oris: Vec<&str> = oris.trim().split(",").collect();
         for (ori_cnt, ori) in oris.into_iter().enumerate() {
@@ -334,19 +334,19 @@ impl RobInfo {
     pub fn rob_name(&self) -> String {
         self.rob_name.clone()
     }
-    pub fn pos_to_zero(&self) -> [f32; 3] {
+    pub fn pos_to_zero(&self) -> [f64; 3] {
         self.pos_for_zero
     }
-    pub fn ori_to_zero(&self) -> [f32; 3] {
+    pub fn ori_to_zero(&self) -> [f64; 3] {
         self.ori_for_zero
     }
-    pub fn min_embed_height(&self) -> f32 {
+    pub fn min_embed_height(&self) -> f64 {
         self.min_embed_height
     }
 }
 
 //Helper function for both cam and rob config to extract xyz coords/rotations from a given string surrounded by "[]" and delimited by ","
-fn pos_ori_parser(line: String) -> Result<[f32; 3], anyhow::Error> {
+fn pos_ori_parser(line: String) -> Result<[f64; 3], anyhow::Error> {
     //Access the string array
     let line_split: Vec<&str> = line.split("[").collect();
 
@@ -354,7 +354,7 @@ fn pos_ori_parser(line: String) -> Result<[f32; 3], anyhow::Error> {
     let vals = line_split[1].replace("]", "");
 
     //Place each value in an actual array
-    let mut out: [f32; 3] = [-0.1, -0.1, -0.1];
+    let mut out: [f64; 3] = [-0.1, -0.1, -0.1];
 
     for (cnt, token) in vals.split(",").enumerate() {
         //Check the count is correct
